fix: rename vector to vector_utils
