+++ src/core/search/CMakeLists.txt
@@ -5,7 +5,7 @@ cur_gen_dir(gen_dir)
 
 find_package(ICU REQUIRED COMPONENTS uc i18n)
 
-add_library(query_parser ast_expr.cc query_driver.cc search.cc indices.cc vector.cc compressed_sorted_set.cc
+add_library(query_parser ast_expr.cc query_driver.cc search.cc indices.cc vector_utils.cc compressed_sorted_set.cc
     ${gen_dir}/parser.cc ${gen_dir}/lexer.cc)
 
 target_link_libraries(query_parser ICU::uc ICU::i18n)
