+++ src/core/search/vector_utils.h
